Fixed authentication logic, using useEffect in a AuthWrapper
