+++ src/components/Home.js
@@ -1,9 +1,13 @@
 import React from "react";
+import LogoutButton from "./LogoutButton";
+import LoginButton from "./LoginButton";
 
 const Home = () => {
   return (
     <div>
       <h1>Home</h1>
+      <LoginButton />
+      <LogoutButton />
     </div>
   );
 };
