+++ src/components/LoginButton.js
@@ -4,24 +4,9 @@ import { useNavigate } from "react-router-dom";
 import axios from "axios";
 
 const LoginButton = () => {
-  const { loginWithRedirect, isAuthenticated, user } = useAuth0();
-  const navigate = useNavigate();
+  const { loginWithRedirect } = useAuth0();
 
   const loginAndRedirect = async () => {
-    if (isAuthenticated && user.logins_count === 0) {
-      axios.post("http://localhost:3000/users/", {
-        firstName: "Charles",
-        lastName: "Lee",
-        userEmail: user.email,
-        username: user.nickname,
-      });
-      // If the user is already logged in and has not logged in before, redirect to onboarding page - change route later"
-      navigate("/");
-    } else if (isAuthenticated && user.logins_count > 0) {
-      // If the user is already logged in, redirect to the home page
-      navigate("/");
-    }
-
     await loginWithRedirect();
   };
