+++ src/index.js
@@ -1,10 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { createRoot } from "react-dom/client";
+import Home from "./components/Home";
 import { Auth0Provider } from "@auth0/auth0-react";
 import "./index.css";
 import App from "./App";
+import AuthWrapper from "./components/AuthWrapper";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 
@@ -12,7 +13,7 @@ root.render(
   <Auth0Provider
     domain={process.env.REACT_APP_DOMAIN}
     clientId={process.env.REACT_APP_CLIENT_ID}
-    redirectUri={window.location.origin}
+    redirectUri={`${window.location.origin}/home`}
     audience={process.env.REACT_APP_AUDIENCE}
     scope="read:current_user update:current_user_metadata"
   >
@@ -21,7 +22,16 @@ root.render(
         {/* Route that provides base app UI */}
         <Route path="/" element={<App />}>
           {/* Route that matches all other paths */}
-          <Route path="*" element={"Nothing here!"} />
+          <Route
+            index
+            path="/home"
+            element={
+              <AuthWrapper>
+                <Home />
+              </AuthWrapper>
+            }
+          />
+          {/* <Route path="*" element={"Nothing here!"} /> */}
         </Route>
       </Routes>
     </BrowserRouter>
